@@ -1,7 +1,5 @@
 package com.example.androiduitesting;
 
-import androidx.appcompat.app.AppCompatActivity;
-
 import android.os.Bundle;
 import android.view.View;
 import android.widget.ArrayAdapter;
@@ -9,6 +7,8 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.LinearLayout;
 import android.widget.ListView;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import java.util.ArrayList;
 
@@ -37,32 +37,24 @@ public class MainActivity extends AppCompatActivity {
         //dataList.addAll(Arrays.asList(cities));
 
         cityAdapter = new ArrayAdapter<>(this, R.layout.content, dataList);
-
+        cityList.setOnItemClickListener((adapterView, view, i, l) -> {
+            startActivity(ShowActivity.newIntent(this, cityAdapter.getItem(i)));
+        });
 
         cityList.setAdapter(cityAdapter);
 
         final Button addButton = findViewById(R.id.button_add);
-        addButton.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                nameField.setVisibility(View.VISIBLE);
-            }
-        });
+        addButton.setOnClickListener(v -> nameField.setVisibility(View.VISIBLE));
 
         final Button confirmButton = findViewById(R.id.button_confirm);
-        confirmButton.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                String cityName = newName.getText().toString();
-                cityAdapter.add(cityName);
-                newName.getText().clear();
-                nameField.setVisibility(View.INVISIBLE);
-            }
+        confirmButton.setOnClickListener(v -> {
+            String cityName = newName.getText().toString();
+            cityAdapter.add(cityName);
+            newName.getText().clear();
+            nameField.setVisibility(View.INVISIBLE);
         });
 
         final Button deleteButton = findViewById(R.id.button_clear);
-        deleteButton.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                cityAdapter.clear();
-            }
-        });
+        deleteButton.setOnClickListener(v -> cityAdapter.clear());
     }
 }
